prepare basics for components
