+++ {{ cookiecutter.project_slug }}/src/code_location_{{ cookiecutter.project_slug }}/code_location_{{ cookiecutter.project_slug }}/defs/dbt/resources/duckdb_path.py
@@ -1,0 +1,5 @@
+from dagster import ConfigurableResource
+
+
+class DuckDBPathResource(ConfigurableResource):
+    file_path: str
