+++ {{ cookiecutter.project_slug }}/src/code_location_{{ cookiecutter.project_slug }}/code_location_{{ cookiecutter.project_slug }}/defs/dbt/resources/sql_asset_keys.py
@@ -1,0 +1,1 @@
+duckdb_bar_warehouse_name = "duckdb_bar"
