+++ {{ cookiecutter.project_slug }}/src/code_location_{{ cookiecutter.project_slug }}/code_location_{{ cookiecutter.project_slug }}/defs/logistics/logistics_example.py
@@ -1,0 +1,13 @@
+from dagster import asset
+from dagster_dbt import get_asset_key_for_model
+
+from ..dbt_assets import unpartitioned_assets
+
+
+@asset(
+    compute_kind="python",
+    group_name="logistics",
+    deps=[get_asset_key_for_model([unpartitioned_assets], "my_first_dbt_model")],
+)
+def logistics_example():
+    return 1
