--- {{ cookiecutter.project_slug }}/src/code_location_{{ cookiecutter.project_slug }}/code_location_{{ cookiecutter.project_slug }}/resources/duckdb_path.py
@@ -1,5 +1,0 @@
-from dagster import ConfigurableResource
-
-
-class DuckDBPathResource(ConfigurableResource):
-    file_path: str
