--- {{ cookiecutter.project_slug }}/src/code_location_{{ cookiecutter.project_slug }}/code_location_{{ cookiecutter.project_slug }}/resources/sql_asset_keys.py
@@ -1,1 +1,0 @@
-duckdb_bar_warehouse_name = "duckdb_bar"
